requirements.txt
# Core dependencies
streamlit>=1.30.0
google-generativeai>=0.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# For sentence similarity (consolidation agent)
sentence-transformers>=2.3.0
scikit-learn>=1.3.0

# Utilities
requests>=2.31.0
